- Refactored the component.tsx page into multiple components and utils - Moved the Header and Footer to the layout.tsx page
