+++ src/app/layout.tsx
@@ -1,7 +1,8 @@
 import { GoogleAnalytics } from "@next/third-parties/google";
 import { Geist, Geist_Mono } from "next/font/google";
 import "./globals.css";
-
+import Header from "../components/Header";
+import Footer from "../components/Footer";
 const geistSans = Geist({
   variable: "--font-geist-sans",
   subsets: ["latin"],
@@ -36,14 +37,16 @@ export default function RootLayout({
         <meta name="apple-mobile-web-app-title" content="Base64 Pro" />
         <link rel="manifest" href="/site.webmanifest" />
       </head>
+      
       <body
         className={`${geistSans.variable} ${geistMono.variable} antialiased`}
       >
+        <Header />
         {children}
-      </body>
-      <GoogleAnalytics gaId="G-S88JXFVVB4" />
+        <Footer />
+        <GoogleAnalytics gaId="G-S88JXFVVB4" />
       {/* eslint-disable-next-line @next/next/no-sync-scripts */}
-      <script
+      <script     
         data-name="BMC-Widget"
         data-cfasync="false"
         src="https://cdnjs.buymeacoffee.com/1.0.0/widget.prod.min.js"
@@ -55,6 +58,8 @@ export default function RootLayout({
         data-x_margin="18"
         data-y_margin="18"
       ></script>
+      </body>
+      
     </html>
   );
 }
